requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiokafka==0.12.0
aiosignal==1.3.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
kafka-python-ng==2.2.3
multidict==6.1.0
packaging==24.2
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sseclient==0.0.27
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
